Se genera movimiento en el avion

+++ src/co/edu/uptc/pojo/Coordinate.java
@@ -1,32 +1,32 @@
 package co.edu.uptc.pojo;
 
 public class Coordinate {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Coordinate() {
         x = 0;
         y = 0;
     }
 
-    public Coordinate(int x, int y) {
+    public Coordinate(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
